Promjena tipa podatka u tabeli Recenzije atribut Ocjena

+++ eBarberShop/UI/ebarbershop_mobile/lib/models/recenzije/recenzije.dart
@@ -6,7 +6,7 @@ part 'recenzije.g.dart';
 class Recenzije {
   int recenzijeId;
   String sadrzaj;
-  int ocjena;
+  double ocjena;
   DateTime datumObjave;
   int korisnikId;
   String? imeKorisnika;
