+++ eBarberShop/eBarberShop.Model/Recenzije.cs
@@ -4,7 +4,7 @@
     {
         public int RecenzijeId { get; set; }
         public string Sadrzaj { get; set; }
-        public int Ocjena { get; set; }
+        public double Ocjena { get; set; }
         public DateTime DatumObjave { get; set; }
         public int KorisnikId { get; set; }
         public Korisnici Korisnik { get; set; }
